Feature: add a search form [WIP]
Adds the (beginnings of) a search form to the common footer
that can be used to trigger a search of our global CSE for
the grad.arizona.edu.

With this commit, the form markup is stripped by drupal
before the footer pages is rendered.

This commit also makes the footer links prettier with some
updates to the AZ Boostrap CSS classes and consolidating
the links into 2 columns

references #1